zavrsnifakspreddipl/requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.0.5
requests>=2.31.0
newsapi-python>=0.2.7
